Unmarshal before passing to opentelemetry-logging-collection service builder
