+++ service/builder.go
@@ -2,9 +2,7 @@ package service
 
 import (
 	"context"
-	"errors"
 	"fmt"
-	"path/filepath"
 
 	"github.com/kardianos/service"
 	"github.com/observiq/stanza/v2/operator/helper/persist"
@@ -15,7 +13,7 @@ import (
 
 type AgentServiceBuilder struct {
 	logger       *zap.SugaredLogger
-	configFile   *string
+	config       *Config
 	pluginDir    *string
 	databaseFile *string
 }
@@ -34,8 +32,8 @@ func (b *AgentServiceBuilder) WithLogger(logger *zap.SugaredLogger) *AgentServic
 	return b
 }
 
-func (b *AgentServiceBuilder) WithConfigFile(configFile string) *AgentServiceBuilder {
-	b.configFile = &configFile
+func (b *AgentServiceBuilder) WithConfigFile(config *Config) *AgentServiceBuilder {
+	b.config = config
 	return b
 }
 
@@ -45,11 +43,6 @@ func (b *AgentServiceBuilder) WithDatabaseFile(datbaseFile string) *AgentService
 }
 
 func (b *AgentServiceBuilder) Build(ctx context.Context) (service.Service, context.Context, error) {
-	// Check to ensure glob file path is not passed in. We want a single file
-	if err := b.validateNonGlobConfig(); err != nil {
-		return nil, context.TODO(), err
-	}
-
 	logAgent, err := b.buildAgent()
 	if err != nil {
 		return nil, context.TODO(), err
@@ -63,29 +56,13 @@ func (b *AgentServiceBuilder) Build(ctx context.Context) (service.Service, conte
 	return newAgentService(ctx, logAgent, persister, persisterShutdownFunc)
 }
 
-// validateNonGlobConfig returns an error if the configFile pointer is a glob.
-// Return nil if the configFile pointer is nil or not a single file path
-func (b *AgentServiceBuilder) validateNonGlobConfig() error {
-	if b.configFile == nil {
-		return nil
-	}
-
-	matches, err := filepath.Glob(*b.configFile)
-	switch {
-	case err != nil:
-		return err
-	case len(matches) > 1:
-		return errors.New("glob config paths not supported")
-	default:
-		return nil
-	}
-}
-
 func (b *AgentServiceBuilder) buildAgent() (*agent.LogAgent, error) {
 	agentBuilder := agent.NewBuilder(b.logger)
 
-	if b.configFile != nil {
-		agentBuilder = agentBuilder.WithConfigFiles([]string{*b.configFile})
+	if b.config != nil {
+		agentBuilder = agentBuilder.WithConfig(&agent.Config{
+			Pipeline: b.config.Pipeline,
+		})
 	}
 
 	if b.pluginDir != nil {
